Separate transaction commiting module
